backend/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart
yt-dlp
browser-cookie3
pathlib
python-dotenv


# ... existing code ... 